Split out parley_core cate
