+++ parley/src/builder.rs
@@ -10,11 +10,11 @@ use super::style::{Brush, StyleProperty, TextStyle, WhiteSpaceCollapse};
 use super::layout::Layout;
 
 use alloc::string::String;
-use core::ops::{Bound, Range, RangeBounds};
+use core::ops::RangeBounds;
 
 use crate::InlineBoxKind;
-use crate::inline_box::InlineBox;
-use crate::resolve::{StyleRun, tree::ItemKind};
+use crate::resolve::tree::ItemKind;
+use parley_core::InlineBox;
 
 /// Builder for constructing a text layout with ranged attributes.
 #[must_use]
@@ -29,8 +29,7 @@ impl<B: Brush> RangedBuilder<'_, B> {
     pub fn push_default<'a>(&mut self, property: impl Into<StyleProperty<'a, B>>) {
         let resolved = self
             .lcx
-            .rcx
-            .resolve_property(self.fcx, &property.into(), self.scale);
+            .resolve_style_property(self.fcx, self.scale, &property.into());
         self.lcx.ranged_style_builder.push_default(resolved);
     }
 
@@ -41,111 +40,19 @@ impl<B: Brush> RangedBuilder<'_, B> {
     ) {
         let resolved = self
             .lcx
-            .rcx
-            .resolve_property(self.fcx, &property.into(), self.scale);
+            .resolve_style_property(self.fcx, self.scale, &property.into());
         self.lcx.ranged_style_builder.push(resolved, range);
     }
 
     pub fn push_inline_box(&mut self, inline_box: InlineBox) {
-        self.lcx.inline_boxes.push(inline_box);
+        self.lcx.core_ctx.push_inline_box(inline_box);
     }
 
     pub fn build_into(self, layout: &mut Layout<B>, text: impl AsRef<str>) {
         // Apply RangedStyleBuilder styles directly to style-table/style-run state.
-        self.lcx
-            .ranged_style_builder
-            .finish(&mut self.lcx.style_table, &mut self.lcx.style_runs);
+        self.lcx.ranged_style_builder.finish(&mut self.lcx.core_ctx);
 
         // Call generic layout builder method
-        self.lcx.build_into(
-            &mut layout.data,
-            self.scale,
-            self.quantize,
-            text.as_ref(),
-            self.fcx,
-        );
-    }
-
-    pub fn build(self, text: impl AsRef<str>) -> Layout<B> {
-        let mut layout = Layout::default();
-        self.build_into(&mut layout, text);
-        layout
-    }
-}
-
-/// Builder for constructing a text layout from a style table and
-/// indexed style runs.
-#[must_use]
-pub struct StyleRunBuilder<'a, B: Brush> {
-    pub(crate) scale: f32,
-    pub(crate) quantize: bool,
-    pub(crate) len: usize,
-    pub(crate) lcx: &'a mut LayoutContext<B>,
-    pub(crate) fcx: &'a mut FontContext,
-    pub(crate) cursor: usize,
-}
-
-impl<B: Brush> StyleRunBuilder<'_, B> {
-    /// Reserves additional capacity for styles and runs.
-    ///
-    /// This is an optional optimization for callers that know counts
-    /// up front; call it before pushing styles and runs to reduce
-    /// reallocations.
-    pub fn reserve(&mut self, additional_styles: usize, additional_runs: usize) {
-        self.lcx.style_table.reserve(additional_styles);
-        self.lcx.style_runs.reserve(additional_runs);
-    }
-
-    /// Adds a fully-specified style to the shared style table and
-    /// returns its index.
-    pub fn push_style<'family, 'settings>(
-        &mut self,
-        style: TextStyle<'family, 'settings, B>,
-    ) -> u16 {
-        let resolved = self
-            .lcx
-            .rcx
-            .resolve_entire_style_set(self.fcx, &style, self.scale);
-        let style_index = self.lcx.style_table.len();
-        assert!(style_index <= u16::MAX as usize, "too many styles");
-        self.lcx.style_table.push(resolved);
-        style_index as u16
-    }
-
-    /// Adds a style run referencing an entry from the style table.
-    ///
-    /// Runs must be contiguous and non-overlapping, and must cover
-    /// `0..text.len()` once all runs have been added.
-    pub fn push_style_run(&mut self, style_index: u16, range: impl RangeBounds<usize>) {
-        let range = resolve_range(range, self.len);
-        assert!(
-            range.start == self.cursor,
-            "StyleRunBuilder expects contiguous non-overlapping runs"
-        );
-        assert!(
-            range.start <= range.end,
-            "StyleRunBuilder expects ordered ranges"
-        );
-        assert!(
-            (style_index as usize) < self.lcx.style_table.len(),
-            "StyleRunBuilder expects style indices that were previously added via push_style"
-        );
-        self.lcx.style_runs.push(StyleRun {
-            style_index,
-            range: range.clone(),
-        });
-        self.cursor = range.end;
-    }
-
-    pub fn push_inline_box(&mut self, inline_box: InlineBox) {
-        self.lcx.inline_boxes.push(inline_box);
-    }
-
-    pub fn build_into(self, layout: &mut Layout<B>, text: impl AsRef<str>) {
-        assert!(
-            self.cursor == self.len,
-            "StyleRunBuilder requires runs that cover the full text"
-        );
         self.lcx.build_into(
             &mut layout.data,
             self.scale,
@@ -173,10 +80,7 @@ pub struct TreeBuilder<'a, B: Brush> {
 
 impl<B: Brush> TreeBuilder<'_, B> {
     pub fn push_style_span(&mut self, style: TextStyle<'_, '_, B>) {
-        let resolved = self
-            .lcx
-            .rcx
-            .resolve_entire_style_set(self.fcx, &style, self.scale);
+        let resolved = self.lcx.resolve_style_set(self.fcx, self.scale, &style);
         self.lcx.tree_style_builder.push_style_span(resolved);
     }
 
@@ -187,11 +91,13 @@ impl<B: Brush> TreeBuilder<'_, B> {
         's: 'iter,
         B: 'iter,
     {
-        self.lcx.tree_style_builder.push_style_modification_span(
-            properties
-                .into_iter()
-                .map(|p| self.lcx.rcx.resolve_property(self.fcx, p, self.scale)),
-        );
+        self.lcx
+            .tree_style_builder
+            .push_style_modification_span(properties.into_iter().map(|p| {
+                self.lcx
+                    .core_ctx
+                    .resolve_style_property(self.fcx, self.scale, p)
+            }));
     }
 
     pub fn pop_style_span(&mut self) {
@@ -213,7 +119,7 @@ impl<B: Brush> TreeBuilder<'_, B> {
 
         // TODO: arrange type better here to factor out the index
         inline_box.index = self.lcx.tree_style_builder.current_text_len();
-        self.lcx.inline_boxes.push(inline_box);
+        self.lcx.core_ctx.push_inline_box(inline_box);
     }
 
     pub fn set_white_space_mode(&mut self, white_space_collapse: WhiteSpaceCollapse) {
@@ -225,10 +131,7 @@ impl<B: Brush> TreeBuilder<'_, B> {
     #[inline]
     pub fn build_into(self, layout: &mut Layout<B>) -> String {
         // Apply TreeStyleBuilder styles to LayoutContext.
-        let text = self
-            .lcx
-            .tree_style_builder
-            .finish(&mut self.lcx.style_table, &mut self.lcx.style_runs);
+        let text = self.lcx.tree_style_builder.finish(&mut self.lcx.core_ctx);
 
         // Call generic layout builder method
         self.lcx
@@ -243,18 +146,4 @@ impl<B: Brush> TreeBuilder<'_, B> {
         let text = self.build_into(&mut layout);
         (layout, text)
     }
-}
-
-fn resolve_range(range: impl RangeBounds<usize>, len: usize) -> Range<usize> {
-    let start = match range.start_bound() {
-        Bound::Unbounded => 0,
-        Bound::Included(n) => *n,
-        Bound::Excluded(n) => *n + 1,
-    };
-    let end = match range.end_bound() {
-        Bound::Unbounded => len,
-        Bound::Included(n) => *n + 1,
-        Bound::Excluded(n) => *n,
-    };
-    start.min(len)..end.min(len)
 }
