+++ parley/src/layout/line.rs
@@ -6,10 +6,10 @@ use crate::layout::data::BreakReason;
 use crate::layout::data::LineData;
 use crate::layout::layout::Layout;
 use crate::layout::run::Run;
-use crate::shape::data::{Glyph, LayoutItemKind};
 use crate::style::Brush;
 use crate::{InlineBox, InlineBoxKind};
 use core::ops::Range;
+use parley_core::{Glyph, LayoutItemKind};
 
 /// Line in a text layout.
 #[derive(Copy, Clone)]
