+++ parley_core/src/shape/cache.rs
