+++ parley_core/src/tests/mod.rs
@@ -1,5 +1,4 @@
 // Copyright 2024 the Parley Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-mod test_builders;
-mod utils;
+mod test_analysis;
